refactor Remove Rails and set-up as Grape only application
It was my intention to see which Ruby framework was more performant i.e. Grape or Rails/Grape.
Having removed much of the bloat from Rails it was comparable to Grape only API however was still slower.
The programming test suggests that performance should be the highest priority therefore the application was refactored to remove Rails.